Add instructions to README and necessary files
Updated the README.md to include instructions on how to convert a WordPress VIP site to a static HTML site. Included copies of 3rd party plugins necessary for the conversion in case the origin doesn't work in the future. Added functions.php with necessary site code to include in your WordPress site.